requirements.txt
openai>=1.12.0
faiss-cpu>=1.7.4
numpy>=1.24.0
python-dotenv>=1.0.0

